feat[web](shoppingList): add product suggestion
feat[web](shoppingList): improve handling kotlin classes in JS via add to class objects toJs method
refactor(core): adjust val names to their file names
refactor: remove unused imports from gradle files

+++ shoppingList/src/commonMain/kotlin/pl/kurczaczkowe/bill/shoppingList/dto/AddToShoppingListParameters.kt
@@ -8,7 +8,7 @@ import kotlin.js.JsExport
 data class AddToShoppingListParameters(
     val shopping_list_id: Long,
     val product_unit: UnitEnum,
-    val product_quantity: Double,
+    val product_quantity: Float,
     val product_name: String,
     val category_id: Long,
 )
